requirements.txt
mcp
google-generativeai
google-genai
python-dotenv
asyncio
typing-extensions
setuptools
wheel
pymupdf
fastmcp
pizzapi
requests
pydantic
ollama